requirement.txt
telethon
argostranslate
pandas
google.generativeai